share log-handler
